### 2024-01-08 #80 @Bsingstad
[Re] Deep Learning for ECG Analysis: Benchmarks and Insights from PTB-XL
**Original article:** N. Strodthoff, P. Wagner, T. Schaeffter, and W. Samek, ‘Deep Learning for ECG Analysis: Benchmarks and Insights from PTB-XL’, IEEE Journal of Biomedical and Health Informatics, vol. 25, no. 5, pp. 1519–1528, May 2021, doi: 10.1109/JBHI.2020.3022989

**Scientific domain:** Time series classification
**Suggested editor:** @ThomasA
