### 2024-10-11 #80 @rougier
@Bsingstad sorry very very long delay, the year has been a bit chaotic.
@gdetor Great, many thanks.  There are plenty of reviewers for deep learning in the board, don't hesitate to ask them at once.
