### 2024-10-23 #80 @gdetor
Dear reviewers @ReScience/reviewers Would anybody be interested in reviewing this work?
