### 2025-02-17 #80 @rougier
@ReScience/reviewers We're looking for reviewers (Time series classification / Python), any takers ?
