### 2025-02-18 #80 @ozancaglayan
One very quick comment as I go through the paper and code @Bsingstad : Why did you modify the fast.ai codebase and maintain a forked one? I dont think this should be necessary and it will avoid people to use the upstream one as it is unlikely that you'll keep your fork synced and maintained.

If its only due to an issue with `tqdm`, progress bar, I'm pretty sure you can find another solution in your notebook to disable progress bars or monkey patch the problematic parts.
