### 2025-02-18 #80 @ozancaglayan
Some more remarks:
- Although Google Colab is seen here as a way to speed up experimentation, it would be also useful to provide a standalone script to run, as otherwise people will need to setup colab and upload the file etc.
- I don't use anaconda at all nowadays, how should I install the dependencies? Could you also create a `requirements.txt` file?
- Can you clear cell outputs in from the notebook file? There are lots of cell output messages that make it hard to see the actual cells when opened in a web browser.
