### 2025-02-23 #80 @ozancaglayan
Before proceeding with the code review, here are some minor remarks and questions based on the paper @Bsingstad

**Typos and stuff:**

- 3.4: Experimental setup and code, first sentence: The open(ly) available? or simply the github repository?
- 3.4: second paragraph: … perfom() function => perform() function?
- 5 Discussion, second paragraph: Interestingly the the …

**Some questions**
Is your newly proposed model, Inception Time, the same as the original one but only using a different toolkit? Are the inception time models in Figure 4 from their implementation or from your new implementation? What was the reason to reimplement it using Tensorflow? How does this new model in Table 4 compare to the actual results in the earlier table/figures? The numbers are provided in a standalone way, making it harder to compare to the previous results in the paper.
