### 2025-03-07 #80 @Bsingstad
**Although Google Colab is seen here as a way to speed up experimentation, it would be also useful to provide a standalone script to run, as otherwise people will need to setup colab and upload the file etc**

You can run the experiment using the standalone script named 'reproduce_results.py' in the code folder

**I don't use anaconda at all nowadays, how should I install the dependencies? Could you also create a `requirements.txt` file?**

you can pip install the file named ecg_env.yml

**Can you clear cell outputs in from the notebook file? There are lots of cell output messages that make it hard to see the actual cells when opened in a web browser.**

Good point! Done
