### 2025-03-07 #80 @Bsingstad
> Before proceeding with the code review, here are some minor remarks and questions based on the paper [@Bsingstad](https://github.com/Bsingstad)
>
> **Typos and stuff:**
>
> * 3.4: Experimental setup and code, first sentence: The open(ly) available? or simply the github repository?
> * 3.4: second paragraph: … perfom() function => perform() function?
> * 5 Discussion, second paragraph: Interestingly the the …

_Thank you! This is now corrected in the manuscript and added it to the repo_

> **Some questions** Is your newly proposed model, Inception Time, the same as the original one but only using a different toolkit? Are the inception time models in Figure 4 from their implementation or from your new implementation? What was the reason to reimplement it using Tensorflow? How does this new model in Table 4 compare to the actual results in the earlier table/figures? The numbers are provided in a standalone way, making it harder to compare to the previous results in the paper.

_The inception Time model is the same (initially I did not know that), but the new moment is that I am optimizing hyperparameters for each task. The original paper used the same config for all tasks. The main purpose with the implementation was to verify that._

_Good point! I will add the results from "my" inception model to the table with results from the original paper_
