### 2025-04-22 #80 @ozancaglayan
> you can pip install the file named ecg_env.yml

@Bsingstad I dont think this works as it's a conda yml file. Can you generate a `requirements.txt` file as well?
