### 2025-05-29 #80 @Bsingstad
> > you can pip install the file named ecg_env.yml
>
> [@Bsingstad](https://github.com/Bsingstad) I dont think this works as it's a conda yml file. Can you generate a `requirements.txt` file as well?

Now its done, @ozancaglayan
