### 2025-06-08 #80 @ozancaglayan
Sorry for the delay again. Looking at this now again.

- Unfortunately I'm struggling to run both the requirements.txt based installation and the PTB-XL notebook. The latter fails due to issues related to keras and tensorflow. I think there's some sort of dependency issues due to the new versions. In fact for data processing none of those deps seem relevant as its just a bunch of `wget`s.
  - Outdated note here https://github.com/Bsingstad/Strodthoff-2021/blob/master/README.md#dependencies, I think you can now remove the note about `fastai`

- Re: requirements.txt based environment creation, I tried with many Python versions, the oldest being 3.9 but still can't install all the dependencies.
- I then tried the `get_datasets.sh` script and that one fails with bunch of 404 errors as well. Are you sure that the data is still out there and accessible?

I can understand that its been slightly more than a year after the initial submission of this paper and the dependencies moved on with potential breaking changes, but at the same time reproducible repos should be made to ensure that they work as expected regardless of when they're used. So I'd recommend to revisit the repo to pick the latest versions of tools that are known to work either through requirements.txt type of installation or the notebooks.
